+++ src/pages/CookieRecordPage/CookieLogText/HintIcon/index.tsx
