+++ src/pages/CookieRecordPage/CookieLogText/index.tsx
